requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distlib==0.3.9
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.7
setuptools==78.1.0
virtualenv==20.30.0
Werkzeug==3.1.3
gunicorn==22.0.0